+++ lib/app/index.ts
@@ -1,9 +1,4 @@
-import type {
-  Renderable,
-  RenderKit,
-  Subscribe,
-  PublishFunction,
-} from '../types'
+import type { Renderable, RenderKit, Subscribe, Publish } from '../types'
 import type { State } from '../state'
 import type { JaxsBus } from '../bus'
 import { render, Root } from '../rendering/templates/root'
@@ -12,7 +7,7 @@ import { startNavigation } from '../navigation/start'
 export class App {
   window: Window
   document: Document
-  publish: PublishFunction
+  publish: Publish<any>
   subscribe: Subscribe
   bus: JaxsBus
   state: State
