+++ lib/rendering/jsx.ts
@@ -1,4 +1,9 @@
-import type { JsxCollection, Props, Template, Renderable } from '../types'
+import type {
+  RenderableCollection,
+  Props,
+  Template,
+  Renderable,
+} from '../types'
 import { Tag } from './templates/tag'
 import { Children } from './templates/children'
 import { ensureJsxChildrenArray } from './templates/children/normalize'
@@ -7,7 +12,7 @@ import { packageJsxAttributes } from './templates/tag/attributes-and-events'
 const jsx = <T>(
   type: string | Template<T>,
   attributes: Props<T>,
-  ...children: JsxCollection
+  ...children: RenderableCollection
 ): Renderable => {
   if (typeof type === 'string') {
     return new Tag(type, attributes, children)
@@ -16,7 +21,10 @@ const jsx = <T>(
   return type(packageJsxAttributes(attributes, children))
 }
 
-jsx.fragment = <T>(attributes: Props<T>, maybeChildren: JsxCollection) => {
+jsx.fragment = <T>(
+  attributes: Props<T>,
+  maybeChildren: RenderableCollection,
+) => {
   const children = ensureJsxChildrenArray(maybeChildren, attributes)
   return new Children(children)
 }
