+++ package.json
@@ -11,7 +11,7 @@
     "e2e:server": "vite dev --config ./cypress/apps/vite.config.js",
     "e2e": "cypress open",
     "e2e:ci": "cypress run",
-    "lint": "prettier . --write"
+    "lint": "prettier . --write && tsc --noEmit"
   },
   "files": [
     "dist"
